refactor(opal): sever app-layer imports for npm publish
Move Popover, Separator, ShadowDiv, and the Section layout primitive
into web/lib/opal/. Old paths in web/src/refresh-components/ and
web/src/layouts/general-layouts.tsx now re-export from opal so the
~160 downstream consumers keep working unchanged.

Drop the @/* path map from web/lib/opal/tsconfig.json — opal source
tree is now self-contained, with zero @/ imports. Required step toward
publishing @onyx-ai/opal to npm.
